refactor: extract linux open command helpers
